Create Hive Query
This file contains the Hive query that was used for our assignment. The Query was then executed through the command prompt (SSH) in the main cluster node that we created in our bucket. When we press the main nodes SSH button, it opens a command prompt in GCP, and we can execute any queries we want here to process through our pipeline.